9.1.3 add internet available check into the provider. If internet is unavailable, it will reject the promises

+++ src/api/provider.js
@@ -1,29 +1,63 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getDestinations() {
-    return this._api.getDestinations();
+    if (isOnline()) {
+      return this._api.getDestinations();
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   getOffers() {
-    return this._api.getOffers();
+    if (isOnline()) {
+      return this._api.getOffers();
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   getPoints() {
-    return this._api.getPoints();
+    if (isOnline()) {
+      return this._api.getPoints();
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   createPoint(point) {
-    return this._api.createPoint(point);
+    if (isOnline()) {
+      return this._api.createPoint(point);
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   updatePoint(id, point) {
-    return this._api.updatePoint(id, point);
+    if (isOnline()) {
+      return this._api.updatePoint(id, point);
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 
   deletePoint(id) {
-    return this._api.deletePoint(id);
+    if (isOnline()) {
+      return this._api.deletePoint(id);
+    }
+
+    // TODO: Write logic if internet is unavailable
+    return Promise.reject(`Offline logic is not implemented`);
   }
 }
